+++ src/public-api.ts
@@ -5,3 +5,4 @@
 export * from "./lib/mat-multi-sort.directive";
 export * from "./lib/mat-multi-sort-table-data-source";
 export * from "./lib/mat-multi-sort-header/mat-multi-sort-header.component";
+export * from "./lib/mat-multi-sort-control/mat-multi-sort-control.component";